Rename main directories and entities
We will use a more traditional terminology, core/domain will contain entities and domain specific ports, core/common will contain more abstract structures
