+++ src/core/common/persistence/options.ts
@@ -1,0 +1,4 @@
+export type FindOptions = {
+  limit?: number;
+  offset?: number;
+};
