+++ src/core/domain/entities/enums/recipe-type.ts
