+++ src/core/domain/entities/enums/role.ts
