+++ src/core/domain/entities/enums/season.ts
